[VetoCouncilPayout] adding more tests to increase coverage

+++ src/GLOW.sol
@@ -275,8 +275,8 @@ contract Glow is ERC20, IGlow {
 
         uint256 head = pointers.head;
         uint256 tail = pointers.tail;
-        uint newHead = head;
-        uint newTail = tail;
+        uint256 newHead = head;
+        uint256 newTail = tail;
 
         //Loop through the unstaked positions until claimableTotal >= amount
         //Tail will also be <= len so no risk of underflow
